+++ app.yaml
@@ -1,2 +1,2 @@
 runtime: python38
-entrypoint: gunicorn -b :$PORT fcreplay.site.main:app
+entrypoint: gunicorn -b :$PORT fcreplay.site.app:app
