--- fcreplay/site/.gitignore
@@ -1,5 +1,0 @@
-venv
-__pycache__
-fcreplay-site.log
-.vscode
-config.json
