+++ fcreplay/site/app.py
@@ -1,4 +1,4 @@
-from fcreplay.site.site.create_app import create_app
-from fcreplay.site.site.site_config import Config
+from fcreplay.site.create_app import create_app
+from fcreplay.site.site_config import Config
 
 app = create_app(Config)
