--- fcreplay/site/app.yaml
@@ -1,1 +1,0 @@
-runtime: python38
