+++ fcreplay/site/database.py
@@ -1,0 +1,2 @@
+from flask_sqlalchemy import SQLAlchemy
+db = SQLAlchemy()
