--- fcreplay/site/requirements.txt
@@ -1,30 +1,0 @@
-Bootstrap-Flask
-Flask
-Flask-SQLAlchemy
-Flask-WTF
-Jinja2
-MarkupSafe
-Pygments
-SQLAlchemy
-WTForms
-Werkzeug
-appdirs
-autopep8
-certifi
-chardet
-click
-dominate
-google-python-cloud-debugger
-gunicorn
-idna
-itsdangerous
-jedi
-parso
-prompt-toolkit
-psycopg2
-pycodestyle
-requests
-toml
-urllib3
-visitor
-wcwidth
